Fixes: #4597 Parse URI path with an endpoint
Adds functionality to LinkGenerator to parse a URI path given a way to
find an endpoint. This is the replacement for various machinications
using the global route collection and `RouteData.Routers` in earlier
versions.

For now I'm just adding a way to do this using Endpoint Name since it's
a pretty low level feature. Endpoint Name is also very direct, so it
feels good for something like this.

I added this to LinkGenerator because I think it feels like the right
thing do, despite the naming conflict. I don't really want to create a
new top-level service for this.